[CHORE] remove legacy rooms module
