+++ backend/app.py
@@ -28,14 +28,14 @@ from plugins.damage_types import load_damage_type
 from autofighter.cards import award_card
 from autofighter.relics import award_relic
 from autofighter.party import Party
-from autofighter.rooms import BattleRoom
-from autofighter.rooms import BossRoom
-from autofighter.rooms import ChatRoom
-from autofighter.rooms import RestRoom
-from autofighter.rooms import ShopRoom
-from autofighter.rooms import _build_foes
-from autofighter.rooms import _scale_stats
-from autofighter.rooms import _serialize
+from autofighter.rooms.battle import BattleRoom
+from autofighter.rooms.boss import BossRoom
+from autofighter.rooms.chat import ChatRoom
+from autofighter.rooms.rest import RestRoom
+from autofighter.rooms.shop import ShopRoom
+from autofighter.rooms.utils import _build_foes
+from autofighter.rooms.utils import _scale_stats
+from autofighter.rooms.utils import _serialize
 from autofighter.stats import Stats
 from autofighter.stats import apply_status_hooks
 from autofighter.gacha import GachaManager
