+++ backend/autofighter/rooms/boss.py
@@ -1,0 +1,12 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+from .battle import BattleRoom
+
+
+@dataclass
+class BossRoom(BattleRoom):
+    """Boss rooms are tougher encounters with high scaling."""
+
+    strength: float = 100.0
