+++ backend/autofighter/rooms/chat.py
@@ -1,0 +1,33 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Any
+
+from ..party import Party
+from ..passives import PassiveRegistry
+
+from . import Room
+from .utils import _serialize
+
+
+@dataclass
+class ChatRoom(Room):
+    """Chat rooms forward a single message to an LLM character."""
+
+    async def resolve(self, party: Party, data: dict[str, Any]) -> dict[str, Any]:
+        registry = PassiveRegistry()
+        for member in party.members:
+            registry.trigger("room_enter", member)
+        message = data.get("message", "")
+        party_data = [_serialize(p) for p in party.members]
+        return {
+            "result": "chat",
+            "message": message,
+            "party": party_data,
+            "gold": party.gold,
+            "relics": party.relics,
+            "cards": party.cards,
+            "rdr": party.rdr,
+            "card": None,
+            "foes": [],
+        }
