+++ backend/tests/test_battle_loot_items.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 import pytest
 
-import autofighter.rooms as rooms_module
+import autofighter.rooms.battle as rooms_module
 
 
 @pytest.fixture()
