+++ backend/tests/test_damage_type_on_action.py
@@ -1,7 +1,7 @@
 import pytest
 
 from autofighter.party import Party
-from autofighter.rooms import BattleRoom
+from autofighter.rooms.battle import BattleRoom
 from autofighter.stats import Stats
 from autofighter.mapgen import MapNode
 from plugins.damage_types._base import DamageTypeBase
