+++ backend/tests/test_relic_rewards.py
@@ -1,5 +1,5 @@
 import pytest
-import autofighter.rooms as rooms_module
+import autofighter.rooms.battle as rooms_module
 from autofighter.mapgen import MapNode
 from autofighter.party import Party
 from autofighter.stats import Stats
