+++ backend/tests/test_shop_room.py
@@ -1,7 +1,7 @@
 import pytest
 
 from autofighter.party import Party
-from autofighter.rooms import ShopRoom
+from autofighter.rooms.shop import ShopRoom
 from autofighter.mapgen import MapNode
 from plugins.players._base import PlayerBase
