+++ backend/tests/test_wind_multi_target.py
@@ -5,7 +5,7 @@ import pytest
 from autofighter.effects import EffectManager
 from autofighter.mapgen import MapNode
 from autofighter.party import Party
-from autofighter.rooms import BattleRoom
+from autofighter.rooms.battle import BattleRoom
 from autofighter.stats import Stats
 from plugins.damage_types.wind import Wind
